Implemented a caching mechanism for file tree nodes
This ensures that only one version of a file node will exist for any given
file. If a new model for that file is created, the cached file node will be
returned. This allows us to refresh file trees without losing existing file
references and/or creating a memory leak.